brute 40

+++ smallPREOI/Day3/pizza/main.cpp
@@ -5,17 +5,17 @@ using namespace std;
 
 const int MAXN = 1007;
 
-vector<int> lengthA(MAXN);
-vector<int> lengthB(MAXN);
-vector<int> areas(MAXN* MAXN);
+vector<long long> lengthA(MAXN);
+vector<long long> lengthB(MAXN);
+vector<long long> areas(MAXN* MAXN);
 
 int main() {
     ios_base::sync_with_stdio(0);
     cin.tie(0);
-    int a, b, n, k;
+    long long a, b, n, k;
     cin >> a >> b >> n >> k;
 
-    int curr = 0, x;
+    long long curr = 0, x;
     for (int i = 0; i < n; i++) {
         cin >> x;
         lengthA[i] = x - curr;
@@ -31,7 +31,7 @@ int main() {
     }
     lengthB[n] = b - x;
 
-    int y = 0;
+    long long y = 0;
     for (int i = 0; i <= n; i++) {
         for (int j = 0; j <= n; j++) {
             areas[y] = lengthA[i] * lengthB[j];
